pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pilot"
version = "1.0.0"
description = "PILOT v2 — CLI pipeline orchestrator"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
include = ["pilot*"]

[tool.setuptools.package-data]
pilot = [
    "defaults/agents/*.md",
    "defaults/templates/**/*",
    "defaults/README.md",
]

[project.scripts]
pilot = "pilot.cli:main"
